JWT refreshing
Im going to try to commit changes after each change so this is the token refreshing functionality for the frontend. Basically, whenever a get request is made from the frontend, if there is a 404 Unauthorized error then it checks for a refresh token. Then, it uses the refresh token to get a new access token, setting it in local storage. a refresh token is required so it doesnt try to get a new access token if the user wasnt logged in the first place. the refreshToken.tsx file is an axios instance to specifically handle this issue.