UI Changes, Quality of Life changes, and bug fixes
Cleaned up Sign Up and Login UI. Removed the Comic and Nuzlocke info from every page except on the Rules page. History page is protected now using the function on a new file 'ProtectedRoute.' Added a title to History page. The rest is a bunch of small changes for clarity/readability/consistency.