Frontend: Game page
Game page now has CRUD functionality for Team to get, create, update, and delete team for the game. A few changes had to be made in the backend like increasing the mas length of Slug field, rejoining ListAPIView and CreateAPIView, and fixing TeamDetail to correctly perform a partial update. A modal is included to confirm the user wants to delete their team for the game. Features pending for Game page is import and exporting json data so the user can save their team data on their computer and import it whenever they want. Everything else needs to be prettier. There is a bug where the site is in a signed in state even though the refresh token has expired.