Fixed signed in state bug
I fixed the bug where the signed in state would remain true even after the refresh token has expired. I fixed this by adding a then and catch handler to the post method in Profile to get a new access token so that if there is another 401 error it clears the local storage. This will lead to the token check in App.tsx to change the signed state to false, fixing the bug.