Dark mode and cleaning up UI
Added dark mode to website and prettified UI more. I've begun learning about testing on the backend. I also rearranged component files on the front end for easier containerization of components. I had to take a break and forgot to upload these changes. Better code to look forward to in the future.